Refactor controllers to namespace under Api::V1 and migrate related specs
