+++ spec/controllers/customers_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe CustomersController, type: :controller do
+RSpec.describe Api::V1::CustomersController, type: :controller do
   include Devise::Test::ControllerHelpers
 
   describe 'GET #index' do
